fix(audio-studio): satisfy GATE-11/format/persistence/name-audit gates
Phase 2 port-in regressed four CI gates; this clears all of them without
raising any budget (the size ratchet only decreases):

- GATE-11: chapter_workbench.py grew to 936 (budget 899). Extract the two
  self-contained helper dialogs (SilenceParamsDialog, AcxResultDialog) into
  the new quill/ui/audio_studio/chapter_workbench_dialogs.py; the Workbench
  module drops to 818 and the dialogs keep their names/classification.
- ruff format: run on the Phase 2 audio_studio source + tests so
  `ruff format --check` passes (18 files reformatted).
- persistence audit: classify the five new audio_studio save_* sites
  (book_prefs/history/library/play_queue/sleep_timer) as "content" in
  _REVIEWED_PERSISTENCE, matching the radio/podcasts store shape.
- accessible-name + dialog-inventory snapshots: regenerated so the two
  moved dialog surfaces and the two new control sites (PlayQueueDialog
  ListBox, SleepTimerDialog SpinCtrl) are registered.
