+++ tests/unit/core/audio_studio/test_audio_studio_history.py
