+++ tests/unit/core/audio_studio/test_book_prefs.py
