+++ tests/unit/ui/audio_studio/test_library_tree.py
