+++ tests/unit/ui/audio_studio/test_play_queue_dialog.py
