+++ tests/unit/ui/audio_studio/test_sleep_timer_dialog.py
