+++ tests/unit/ui/fixtures/accessible_name_inventory.json
@@ -108,8 +108,8 @@
   "quill/ui/audio_studio/chapter_workbench.py::ChapterWorkbenchDialog.__init__::wx.ListBox": "named",
   "quill/ui/audio_studio/chapter_workbench.py::ChapterWorkbenchDialog.__init__::wx.TextCtrl": "named",
   "quill/ui/audio_studio/chapter_workbench.py::ChapterWorkbenchDialog._tag_field::wx.TextCtrl": "named",
-  "quill/ui/audio_studio/chapter_workbench.py::SilenceParamsDialog.__init__::wx.SpinCtrlDouble": "named",
-  "quill/ui/audio_studio/chapter_workbench.py::SilenceParamsDialog.__init__::wx.SpinCtrlDouble#2": "named",
+  "quill/ui/audio_studio/chapter_workbench_dialogs.py::SilenceParamsDialog.__init__::wx.SpinCtrlDouble": "named",
+  "quill/ui/audio_studio/chapter_workbench_dialogs.py::SilenceParamsDialog.__init__::wx.SpinCtrlDouble#2": "named",
   "quill/ui/audio_studio/feed_dialog.py::FolderFeedDialog.__init__::wx.ListBox": "named",
   "quill/ui/audio_studio/feed_dialog.py::FolderFeedDialog._field::wx.TextCtrl": "named",
   "quill/ui/audio_studio/pages_audio.py::AudioSourcePage.__init__::wx.ComboBox": "named",
@@ -145,6 +145,7 @@
   "quill/ui/audio_studio/pages_shared.py::BookPage._field::wx.TextCtrl": "named",
   "quill/ui/audio_studio/pages_shared.py::BookPage._ms_spin::wx.SpinCtrl": "named",
   "quill/ui/audio_studio/pages_shared.py::SummaryPage.__init__::wx.TextCtrl": "named",
+  "quill/ui/audio_studio/play_queue_dialog.py::PlayQueueDialog.__init__::wx.ListBox": "named",
   "quill/ui/audio_studio/player_panel.py::PlayerPanel.__init__::wx.Choice": "named",
   "quill/ui/audio_studio/player_panel.py::PlayerPanel.__init__::wx.Slider": "named",
   "quill/ui/audio_studio/player_panel.py::PlayerPanel.__init__::wx.Slider#2": "named",
@@ -155,6 +156,7 @@
   "quill/ui/audio_studio/publish_dialog.py::PublishDialog.__init__::wx.TextCtrl#2": "named",
   "quill/ui/audio_studio/publish_dialog.py::PublishDialog.__init__::wx.TextCtrl#3": "named",
   "quill/ui/audio_studio/publish_dialog.py::PublishDialog._field::wx.TextCtrl": "named",
+  "quill/ui/audio_studio/sleep_timer_dialog.py::SleepTimerDialog.__init__::wx.SpinCtrl": "named",
   "quill/ui/audiobook_chapter_editor_dialog.py::ChapterEditorDialog.__init__::wx.ListBox": "named",
   "quill/ui/audiobook_chapter_editor_dialog.py::ChapterEditorDialog.__init__::wx.TextCtrl": "named",
   "quill/ui/batch_wizard_pages.py::_FolderPage.__init__::wx.DirPickerCtrl": "named",
