+++ tests/unit/ui/fixtures/dialog_inventory.json
@@ -51,14 +51,14 @@
   "quill/ui/assistant_tools.py::RunPythonDialog.__init__::wx.Dialog": "hardened_custom",
   "quill/ui/assistant_tools.py::SearchableModelPickerDialog.__init__::wx.Dialog": "hardened_custom",
   "quill/ui/assistant_tools.py::WritingAssistantDialog.__init__::wx.Dialog": "hardened_custom",
-  "quill/ui/audio_studio/chapter_workbench.py::AcxResultDialog::subclass:wx.Dialog": "hardened_custom",
   "quill/ui/audio_studio/chapter_workbench.py::ChapterWorkbenchDialog._on_export::wx.FileDialog": "native",
   "quill/ui/audio_studio/chapter_workbench.py::ChapterWorkbenchDialog._on_export::wx.SingleChoiceDialog": "native",
   "quill/ui/audio_studio/chapter_workbench.py::ChapterWorkbenchDialog._on_import::wx.FileDialog": "native",
   "quill/ui/audio_studio/chapter_workbench.py::ChapterWorkbenchDialog._on_save_as::wx.FileDialog": "native",
   "quill/ui/audio_studio/chapter_workbench.py::ChapterWorkbenchDialog._on_split_into_files::wx.DirDialog": "native",
   "quill/ui/audio_studio/chapter_workbench.py::ChapterWorkbenchDialog::subclass:wx.Dialog": "hardened_custom",
-  "quill/ui/audio_studio/chapter_workbench.py::SilenceParamsDialog::subclass:wx.Dialog": "hardened_custom",
+  "quill/ui/audio_studio/chapter_workbench_dialogs.py::AcxResultDialog::subclass:wx.Dialog": "hardened_custom",
+  "quill/ui/audio_studio/chapter_workbench_dialogs.py::SilenceParamsDialog::subclass:wx.Dialog": "hardened_custom",
   "quill/ui/audio_studio/feed_dialog.py::FolderFeedDialog::subclass:wx.Dialog": "hardened_custom",
   "quill/ui/audio_studio/library_tree.py::LibraryTreeActions.delete_book::wx.MessageDialog": "native",
   "quill/ui/audio_studio/library_tree.py::LibraryTreeActions.move_to_folder::wx.SingleChoiceDialog": "native",
